refactor : 유저 프로필 이미지 가져오는 메서드를 UserActivity로 이동
