+++ hifi_Shopping/app/src/main/java/com/hifi/hifi_shopping/user/MyPageFragment.kt
@@ -2,14 +2,11 @@ package com.hifi.hifi_shopping.user
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.ViewModelProvider
-import com.google.firebase.database.FirebaseDatabase
-import com.hifi.hifi_shopping.MainActivity
 import com.hifi.hifi_shopping.R
 import com.hifi.hifi_shopping.buy.BuyActivity
 import com.hifi.hifi_shopping.category.CategoryActivity
@@ -17,15 +14,12 @@ import com.hifi.hifi_shopping.databinding.FragmentMyPageBinding
 import com.hifi.hifi_shopping.recommend.RecommendActivity
 import com.hifi.hifi_shopping.search.SearchActivity
 import com.hifi.hifi_shopping.subscribe.SubscribeActivity
-import com.hifi.hifi_shopping.user.model.PointDataClass
-import com.hifi.hifi_shopping.user.model.UserDataClass
 import com.hifi.hifi_shopping.user.vm.OrderViewModel
 import com.hifi.hifi_shopping.user.vm.PointViewModel
 import com.hifi.hifi_shopping.user.vm.ProductViewModel
 import com.hifi.hifi_shopping.user.vm.ReviewViewModel
 import com.hifi.hifi_shopping.user.vm.SubscribeViewModel
 import com.hifi.hifi_shopping.user.vm.UserCouponViewModel
-import java.util.UUID
 
 class MyPageFragment : Fragment() {
 
@@ -135,6 +129,10 @@ class MyPageFragment : Fragment() {
                     }
                     true
                 }
+            }
+
+            myPageUserImg.run {
+                userActivity.getUserProfileImg(userTemp,this)
             }
 
             myPageBtnToEditUser.setOnClickListener {
